requirements.txt
pdf2image==1.16.0
Pillow==8.2.0
pytesseract==0.3.7
